New version of electrical diagram.
Version 10.
Changes:
Changed a lot in this version. New and better PLC connections, also improved the visibility of a lot of connections in the whole project. Also for better overview I did plsit the drive pages in 2 pages. One page for the Power Module wiring, and also one for the Control Module wiring.